test: make the suite pass on MySQL, the database it ships on
Nine tests passed on SQLite and failed on MySQL because SQLite enforces
neither enums nor unsigned columns:

- four wrote payment_method='cod' into an enum that has held only
  'online' since 2026-06-25. The code under test decides gateway-vs-manual
  by a captured PaymentIntent, never by payment_method, so they now model
  a refund owed outside the gateway as an online order with no intent —
  which is what such an order is. PSW-06 loses its COD leg.
- two seeded negative group_bv_daily rows into unsigned columns. The
  reversal engine absorbs down to zero and carries the rest as a debt, so
  that state cannot exist; replaced with the realistic boundary (every BV
  day forfeited → exactly zero) and the two-range total. The service
  comment that repeated the false premise is corrected; the clamp stays.
- one wrote users.status='pending_kyc', a value the enum never had; the
  hold condition is status !== 'active', so 'pending' is the fixture.
- one asserted a raw SUM() as int; MySQL returns a string.
- one asserted an ON DELETE CASCADE with SET FOREIGN_KEY_CHECKS=0 in
  force, which suspends cascades on MySQL but not on SQLite; checks are
  re-enabled for the run under test.

Full suite on MySQL: 2046 passed, 1 skipped, 0 failed.

+++ app/tests/Modules/Commerce/CheckoutZeroCashFloorTest.php
@@ -148,7 +148,7 @@ it('CZC-02: a sub-₹1 payable is taken to ₹1 by applying less credit; the res
 
     // The prepayment posted matches the ₹1 actually payable.
     $tx = LedgerTx::where('idempotency_key', 'order.placed:'.$order->id)->sole();
-    expect(DB::table('ledger_entries')->where('ledger_tx_id', $tx->id)->where('side', 'debit')->sum('amount_paise'))->toBe(100);
+    expect((int) DB::table('ledger_entries')->where('ledger_tx_id', $tx->id)->where('side', 'debit')->sum('amount_paise'))->toBe(100);
 });
 
 it('CZC-04: a fully credit-settled order places at ₹0 with no prepayment, confirms without a gateway, and ships', function () {
